+++ app/javascript/components/countdown_timer.js
@@ -3,9 +3,6 @@ export const countdownTimer = () => {
     const ready = document.getElementById("ready");
     const time = document.getElementById("time");
     const newRoundBtn = document.getElementById("new_round");
-    const countdown = document.getElementById("countdown");
-    const overlay = document.getElementById("overlay")
-    const revealImposter = document.getElementById("impostor-reveal")
     
     const showTimer = () => {
         setTimeout(() => {
@@ -34,25 +31,5 @@ export const countdownTimer = () => {
         countdownStart();
     }
     
-    if(!!overlay){
-        const revealingImposter = () => {
-            const timer = setInterval(() => {
-                countdown.innerText -= 1
-                if((parseInt(countdown.innerText)) === 0){
-                    console.log("hello")
-                    clearInterval(timer);
-                    overlay.classList.toggle("active")
-                    setTimeout(() => {
-                        revealImposter.classList.toggle("d-none")
-                    },2000)
-                }
-            }, 1000)
-        }
-
-    }
-
     if(!!ready){ showReady(); }
-    if(!!countdown) {
-        $( "#countdown" ).load(window.location.href + " #countdown" );
-    }
 }
