+++ app/javascript/packs/application.js
@@ -35,6 +35,7 @@ import { countdownTimer } from "../components/countdown_timer";
 import { autoFocus } from "../components/auto_focus";
 import { updateDiv } from "../components/auto_refresh";
 import { initResultCable } from "../channels/result_channel";
+import { showImposter } from "../components/show_imposter";
 
 
 // next 3 lines is from https://fontsource.org/fonts
@@ -54,4 +55,5 @@ document.addEventListener('turbolinks:load', () => {
   autoFocus();
   updateDiv();
   captureResults();
+  showImposter();
 });
