+++ app/views/results/_result_total_votes.html.erb
@@ -1,0 +1,1 @@
+<h4><%= total_votes %></h4>
